% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{asp_palette}
\alias{asp_palette}
\title{A palette generator for the Sahel Adaptive Social Protection RCT}
\source{
Adapted from
\url{https://github.com/karthik/wesanderson/blob/master/R/colors.R}
}
\usage{
asp_palette(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of desired palette. Current choices are:
\code{IHEID}, \code{Centres}, and \code{SDGs}.}

\item{n}{Number of colors desired. If omitted, uses all colors.}

\item{type}{Either "continuous" or "discrete". Use continuous if you want
to automatically interpolate between colors.}
}
\value{
A vector of colors.
}
\description{
These are a few color palettes useful for the production of graphs for the
Sahel ASP RCT program.
This function calls one of two official palettes in
\code{\link{asp_palette}}: one dark theme and one light theme.
}
\examples{
asp_palette("Dark")
asp_palette("Light")

# If you need more colors than normally found in a palette, you
# can use a continuous palette to interpolate between existing
# colors
pal <- asp_palette(21, name = "Dark", type = "continuous")
image(volcano, col = pal)
}
\keyword{colors}
