% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{asp_palettes}
\alias{asp_palettes}
\title{Complete list of palettes}
\format{
An object of class \code{list} of length 2.
}
\usage{
asp_palettes
}
\description{
Use \code{\link{asp_palette}} to construct palettes of desired length.
}
\keyword{datasets}
