% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collate_models.R
\name{collate_models}
\alias{collate_models}
\title{Collate regression models into a standardised tibble}
\usage{
collate_models(models)
}
\arguments{
\item{models}{list of regression models supported by \code{{broom}}}
}
\value{
A tibble with statistics for multiple models.
}
\description{
Collate regression models into a standardised tibble
}
\examples{
# Use the  R Swiss data for examples with a random treatment
set.seed(1234)
swiss <- dplyr::bind_cols(swiss,
                           treatment = rbinom(seq(nrow(swiss)),1, 0.5)) |>
         dplyr::mutate(Education = Education + 0.15 * treatment)

reg1 <- lm(formula = Education ~ treatment, swiss)
reg2 <- lm(formula = Education ~ treatment + Catholic, swiss)
reg3 <- lm(formula = Education ~ treatment * Catholic, swiss)

models <- list(reg1,
               reg2,
               reg3)
collate_models(models = models)
}
