% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggflatviolin.R
\name{geom_flat_violin_l}
\alias{geom_flat_violin_l}
\title{Flat (or half) vertical violin plot geom. Left side.}
\source{
Adapted from the following \href{https://stackoverflow.com/questions/52034747/plot-only-one-side-half-of-the-violin-plot}{S-O thread}.
}
\usage{
geom_flat_violin_l(
  mapping = NULL,
  data = NULL,
  stat = "ydensity",
  position = "dodge",
  trim = TRUE,
  scale = "area",
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{ggplot2::aes()}. If
specified and \code{inherit.aes = TRUE} (the default), it is combined with the
default mapping at the top level of the plot. You must supply mapping if
there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three options:
If NULL, the default, the data is inherited from the plot data as specified
in the call to ggplot().
A data.frame, or other object, will override the plot data. All objects will
be fortified to produce a data frame. See fortify() for which variables will
be created.
A function will be called with a single argument, the plot data. The return
value must be a data.frame, and will be used as the layer data. A function
can be created from a formula (e.g. ~ head(.x, 10)).}

\item{stat}{Use to override the default connection between \code{geom_violin()}
and \code{stat_ydensity()}.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{trim}{If \code{TRUE} (default), trim the tails of the violins to the range
of the data. If \code{FALSE}, don't trim the tails.}

\item{scale}{if "area" (default), all violins have the same area
(before trimming the tails). If "count", areas are scaled proportionally to
the number of observations. If "width", all violins have the same maximum
width.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes. It can also be a
named logical vector to finely select the aesthetics to display.}

\item{inherit.aes}{Use to override the default connection between
\code{geom_violin()} and \code{stat_ydensity()}.}

\item{...}{Other arguments passed on to \code{layer()}. These are often
aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{linewidth = 3}. They may also be parameters to
the paired geom/stat.}
}
\value{
A half violin vertical geom.
}
\description{
\code{geom_flat_violin_l()} plots half of a violin plot created by
\code{ggplot2::geom_violin()}. This function plots a vertical half of a violin
plot. For the horizontal version of the function, check out
\code{geom_flat_violin_h()}.
}
\examples{
df = data.frame(val = c(), group = c())
for(i in 1:5){
  offset = i - 3
  df = rbind(df,
             data.frame(val = rnorm(n = 50, mean = 0 - offset), group = i)
  )
}
ggplot2::ggplot(df, ggplot2::aes(as.factor(group), val)) +
   geom_flat_violin_l(fill = "lightblue") +
   themeaspdark()
}
