% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sahel_ggboxplot.R
\name{sahel_ggboxplot}
\alias{sahel_ggboxplot}
\title{Individual Boxplot Plots per Country}
\usage{
sahel_ggboxplot(
  data,
  variable,
  countries,
  title = deparse(substitute(variable)),
  subtitle = "Boxplot Plot",
  caption = NULL,
  xtitle = NULL,
  ytitle = NULL,
  scale = NULL,
  theme = themeaspdark
)
}
\arguments{
\item{data}{A data frame}

\item{variable}{Variable name to plot. Continuous or count.}

\item{countries}{Variable name to plot. Continuous or count.}

\item{title}{Plot title}

\item{subtitle}{Plot subtitle}

\item{caption}{Plot caption}

\item{xtitle}{Plot x-axis title}

\item{ytitle}{Plot y-axis title}

\item{scale}{Scaling function from \code{{ggplot2}}}

\item{theme}{Theme applied}
}
\value{
A ggplot object of boxplot style.
}
\description{
Individual Boxplot Plots per Country
}
\examples{
library(patchwork)
sahel_ggboxplot(
  data = sahel_sim,
  variable = hh_income,
  countries = country_names,
  title = "Household Income",
  scale = ggplot2::scale_x_log10
) + sahel_ggboxplot(
  data = sahel_sim,
  variable = hh_consumption,
  countries = country_names,
  title = "Household Consumption",
  scale = ggplot2::scale_x_log10
)
}
