% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sahel_gghalfviolin.R
\name{sahel_gghalfviolin}
\alias{sahel_gghalfviolin}
\title{High-Level Wrapper Half Violin Plots per Country}
\usage{
sahel_gghalfviolin(
  data,
  variable,
  countries,
  orientation = "Vertical-Left",
  alpha_boxplot = 0.5,
  alpha_violin = 0.8,
  width_boxplot = 0.5,
  width_violin = 0.8,
  title = deparse(substitute(variable)),
  subtitle = "Half-Violin Plot",
  caption = NULL,
  xtitle = NULL,
  ytitle = NULL,
  scale = ggplot2::geom_blank(),
  xlim = NULL,
  ylim = NULL,
  add_boxplot = TRUE,
  theme = themeaspdark
)
}
\arguments{
\item{data}{A data frame}

\item{variable}{Variable name to plot. Continuous or count.}

\item{countries}{Variable name to plot. Continuous or count.}

\item{orientation}{Orientation of the graph}

\item{alpha_boxplot}{Transparency for the boxplot element.}

\item{alpha_violin}{Transparency for the violin element.}

\item{width_boxplot}{Width for the boxplot element.}

\item{width_violin}{Width for the violin element.}

\item{title}{Plot title}

\item{subtitle}{Plot subtitle}

\item{caption}{Plot caption}

\item{xtitle}{Plot x-axis title}

\item{ytitle}{Plot y-axis title}

\item{scale}{Scaling function from \code{{ggplot2}}}

\item{xlim}{Define plot limit on x-axis. Supply a vector of bounds, \code{c(325, 500)}.}

\item{ylim}{Define plot limit on y-axis. Supply a vector of bounds, \code{c(325, 500)}.}

\item{add_boxplot}{\code{TRUE} by default. Add a boxplot inside the violin plots?}

\item{theme}{Theme applied}
}
\value{
A ggplot object of half violin style.
}
\description{
This function is a convenience wrapper for the violin plots within the Sahel
ASP plot. If you want a custom plot, check out the \code{geom_flat_violin_l()}
and \code{geom_flat_violin_r()} functions in this package.
}
\examples{
library(patchwork)

# Vertical Left
sahel_gghalfviolin(
  data = sahel_sim,
  variable = hh_income,
  countries = country_names,
  title = "Household Income",
  scale = ggplot2::scale_y_log10()
) + sahel_gghalfviolin(
  data = sahel_sim,
  variable = hh_consumption,
  countries = country_names,
  title = "Household Consumption",
  scale = ggplot2::scale_y_log10()
)

# Vertical Right
sahel_gghalfviolin(
  data = sahel_sim,
  variable = hh_income,
  countries = country_names,
  orientation = "Vertical-Right",
  title = "Household Income",
  scale = ggplot2::scale_y_log10()
) + sahel_gghalfviolin(
  data = sahel_sim,
  variable = hh_consumption,
  countries = country_names,
  orientation = "Vertical-Right",
  title = "Household Consumption",
  scale = ggplot2::scale_y_log10()
)

# Horizontal
sahel_gghalfviolin(
  data = sahel_sim,
  variable = hh_income,
  countries = country_names,
  orientation = "Horizontal",
  title = "Household Income",
  scale = ggplot2::scale_y_log10()
) + sahel_gghalfviolin(
  data = sahel_sim,
  variable = hh_consumption,
  countries = country_names,
  orientation = "Horizontal",
  title = "Household Consumption",
  scale = ggplot2::scale_y_log10()
)
}
