% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sahel_ggviolin.R
\name{sahel_ggviolin}
\alias{sahel_ggviolin}
\title{Individual Violin Plots per Country}
\usage{
sahel_ggviolin(
  data,
  variable,
  countries,
  alpha_boxplot = 0.5,
  alpha_violin = 0.8,
  width_boxplot = 0.5,
  width_violin = 0.8,
  title = deparse(substitute(variable)),
  subtitle = "Violin Plot",
  caption = NULL,
  xtitle = NULL,
  ytitle = NULL,
  scale = NULL,
  add_boxplot = TRUE,
  theme = themeaspdark
)
}
\arguments{
\item{data}{A data frame}

\item{variable}{Variable name to plot. Continuous or count.}

\item{countries}{Variable name to plot. Continuous or count.}

\item{alpha_boxplot}{Transparency for the boxplot element.}

\item{alpha_violin}{Transparency for the violin element.}

\item{width_boxplot}{Width for the boxplot element.}

\item{width_violin}{Width for the violin element.}

\item{title}{Plot title}

\item{subtitle}{Plot subtitle}

\item{caption}{Plot caption}

\item{xtitle}{Plot x-axis title}

\item{ytitle}{Plot y-axis title}

\item{scale}{Scaling function from \code{{ggplot2}}}

\item{add_boxplot}{\code{TRUE} by default. Add a boxplot inside the violin plots?}

\item{theme}{Theme applied}
}
\value{
A ggplot object of violin style.
}
\description{
Individual Violin Plots per Country
}
\examples{
library(patchwork)
sahel_ggviolin(
  data = sahel_sim,
  variable = hh_income,
  countries = country_names,
  title = "Household Income",
  scale = ggplot2::scale_x_log10
) + sahel_ggviolin(
  data = sahel_sim,
  variable = hh_consumption,
  countries = country_names,
  title = "Household Consumption",
  scale = ggplot2::scale_x_log10
)
}
