% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sahel_utilities.R
\name{sahel_prep_gg}
\alias{sahel_prep_gg}
\title{Sahel Prepare ggplot function}
\usage{
sahel_prep_gg(data, key = "variable", value = "value", ...)
}
\arguments{
\item{data}{Dataframe to pivot.}

\item{key}{String. Name of the variable to pivot on.}

\item{value}{String. Name of the value variable.}

\item{...}{Additional agruments passed to \code{tidyr::gather()} such as variables
to pivot or \code{tidyselect} expressions}
}
\value{
A plotting ready tibble.
}
\description{
This function prepares standard ASP data from Stata into a format ready for
plotting with one of the included plotting functions.
}
\examples{
sahel_sim_long_monetary <- tidyr::gather(sahel_sim, key = "variable",
                                         value = "value", hh_income,
                                         hh_consumption)
sahel_sim_long_count <- tidyr::gather(sahel_sim, key = "variable",
                                       value = "value", children_nr,
                                       animals_nr, yr_educ_hhmean)
sahel_sim_long_binary <- tidyr::gather(sahel_sim, key = "variable",
                                       value = "value", has_children,
                                       owns_animals, female_head)

dplyr::summarize(sahel_sim_long_monetary)
dplyr::summarize(sahel_sim_long_count)
dplyr::summarize(sahel_sim_long_binary)
}
