% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sahel_utilities.R
\name{sahel_prep_gg_binary}
\alias{sahel_prep_gg_binary}
\title{Prepare Sahel binary variables for plotting.}
\usage{
sahel_prep_gg_binary(data, ..., alpha = 0.05)
}
\arguments{
\item{data}{A data frame, data frame extension (e.g. a tibble), or a
lazy data frame (e.g. from dbplyr or dtplyr)}

\item{...}{Passed to group_by(), variables or computations to group by.
Computations are always done on the ungrouped data frame. To perform
computations on the grouped data, you need to use a separate mutate()
step before the group_by(). Computations are not allowed in nest_by().
In ungroup(), variables to remove from the grouping.}

\item{alpha}{Signifiicance level for confidence interval}
}
\value{
A dataframe ready for plotting. Computes the
}
\description{
Prepare Sahel binary variables for plotting.
}
\examples{
sahel_sim_long_binary <- tidyr::gather(sahel_sim, key = "variable",
                                       value = "value", has_children,
                                       owns_animals, female_head)
sahel_prep_gg_binary(sahel_sim_long_binary, country_names, variable,
                       alpha = 0.05)
}
