% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sahel_utilities.R
\name{sahel_prep_ggcoefs}
\alias{sahel_prep_ggcoefs}
\title{Sahel Prepare data for \code{sahel_ggcoefs()}}
\usage{
sahel_prep_ggcoefs(
  data,
  variables = c("consum_2_day_eq_ppp", "consum_2_day_ppp", "food_2_day_eq_ppp",
    "food_2_day_ppp", "food_2_day_g_ppp", "food_2_day_g_eq_ppp", "food_2_g_d",
    "FIES_rvrs_raw"),
  var_labels = c("Daily Cons.\\n Ad. Equiv.", "Daily Consumption",
    "Daily Food Cons.\\n Ad. Equiv", "Daily Food Cons.",
    "Gifted Daily Food\\n Cons. Ad. Equiv", "Gifted Daily\\n Food Cons.",
    "Consumed Gifted\\n Food {0,1}", "Food Security"),
  model_number = 1
)
}
\arguments{
\item{data}{A data frame filled with regression output from Stata.}

\item{variables}{A character vector of variable names to plot.}

\item{var_labels}{A chacater vector of corresponding variable names.}

\item{model_number}{Some dataframes have a mht_family variable indicating the
model type number. Select the model you want to plot by passing it the
corresponding number.}
}
\value{
A dataframe ready for barchart plotting.
}
\description{
This function prepares the coefficient information saved post regression
workflow in Stata and makes them ready for plotting.
}
\examples{
# Don't forget to mount the data if using the actual data
# prodregs <-
#  haven::read_dta(r"(U:\fu2_MRT\05_Regstats\fu2_MRT_regstats_hh_prod.dta)")

# Example with simulated data
set.seed(1234)
prodregs <- fabricatr::fabricate(N = 8,
                                 var_name = c("consum_2_day_eq_ppp",
                                               "consum_2_day_ppp",
                                               "food_2_day_eq_ppp",
                                               "food_2_day_ppp",
                                               "food_2_day_g_ppp",
                                               "food_2_day_g_eq_ppp",
                                               "food_2_g_d",
                                               "FIES_rvrs_raw"),
                                 b0 = 0,
                                 b1 = rnorm(N, 0, 1),
                                 b2 = rnorm(N, 0, 1),
                                 b3 = rnorm(N, 0, 1),
                                 avg0 = rnorm(N, 7, 1),
                                 avg1 = rnorm(N, 8, 1),
                                 avg2 = rnorm(N, 9, 1),
                                 avg3 = rnorm(N, 10, 1),
                                 se0 = rnorm(N, 0, 1),
                                 se1 = rnorm(N, 0, 1),
                                 se2 = rnorm(N, 0, 1),
                                 se3 = rnorm(N, 0, 1),
                                 ci95_0 = rnorm(N, 2, 1),
                                 ci95_1 = rnorm(N, 2, 1),
                                 ci95_2 = rnorm(N, 2, 1),
                                 ci95_3 = rnorm(N, 2, 1),
                                 p0 = rnorm(N, 0.8, 0.5),
                                 p1 = rnorm(N, 0.8, 0.5),
                                 p2 = rnorm(N, 0.8, 0.5),
                                 p3 = rnorm(N, 0.8, 0.5),
                                 mht_family = 1) |>
  dplyr::select(-ID)

 sahel_prep_ggcoefs(prodregs)
}
