% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sahel_sim}
\alias{sahel_sim}
\title{\code{sahel_sim} dataset}
\format{
An object of class \code{data.frame} with 7800 rows and 17 columns.
}
\usage{
sahel_sim
}
\description{
Some simulated data for examples in the package
}
\keyword{sahel_sim}
